update AppleTree class - lbsApples is represented by units in Crop class

+++ final-project/Corn.cpp
@@ -20,7 +20,7 @@ double Corn::getBushelsPerAcre() const
     return BUSHELS_PER_ACRE;
 }
 
-double Corn::calcTotalBushels() const
+int Corn::calcTotalBushels() const
 {
     return floor ( Crop::getTotalUnits() * getBushelsPerAcre() );
 }
